+++ backend/src/main/java/me/fairygel/fbook/service/GenreService.java
@@ -1,41 +1,59 @@
 package me.fairygel.fbook.service;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
+import me.fairygel.fbook.dto.genre.GenreDTO;
+import me.fairygel.fbook.dto.genre.GenreIndexViewDTO;
 import me.fairygel.fbook.entity.Genre;
+import me.fairygel.fbook.util.mapper.GenreMapper;
 import me.fairygel.fbook.repository.GenreCrudRepository;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.HashSet;
+import java.util.Set;
 
 @Service
 @AllArgsConstructor
 public class GenreService {
     private final GenreCrudRepository genreCrudRepository;
+    private final GenreMapper mapper;
 
-    public void create(Genre genre) {
+    private static final String NO_GENRE = "No genre with id = ";
+
+    public void create(GenreDTO genreDTO) {
+        Genre genre = mapper.genreDtoToGenre(genreDTO);
         genreCrudRepository.save(genre);
     }
 
-    public Genre read(Long id) {
-        return genreCrudRepository
-                .findById(id).orElse(
-                        genreCrudRepository.findById(0L)
-                                .orElseThrow(IllegalStateException::new)
-                );
+    public GenreDTO read(Long id) {
+        if (id == 0L) throw new EntityNotFoundException(NO_GENRE + id);
+
+        Genre genre = genreCrudRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(NO_GENRE + id));
+        return mapper.genreToGenreDto(genre);
     }
 
-    public Genre update(Long id, Genre genre) {
-        return genreCrudRepository.updateById(id, genre).orElseThrow(IllegalAccessError::new);
+    public GenreDTO update(Long id, GenreDTO genreDTO) {
+        if (id == 0L) throw new EntityNotFoundException(NO_GENRE + id);
+
+        Genre genre = mapper.genreDtoToGenre(genreDTO);
+
+        Genre updatedGenre = genreCrudRepository.updateById(id, genre)
+                .orElseThrow(() -> new EntityNotFoundException(NO_GENRE + id));
+
+        return mapper.genreToGenreDto(updatedGenre);
     }
 
     public void delete(Long id) {
+        if (id == 0L) throw new EntityNotFoundException(NO_GENRE + id);
+
         genreCrudRepository.deleteById(id);
     }
 
-    public List<Genre> index() {
-        List<Genre> genres = new ArrayList<>();
-        genreCrudRepository.findAll().forEach(genres::add);
-        return genres;
+    public Set<GenreIndexViewDTO> index() {
+        Set<Genre> genres = new HashSet<>();
+        genreCrudRepository.findAll().forEach(g -> {if (g.getId() != 0L) genres.add(g);});
+
+        return mapper.genresToIndex(genres);
     }
 }
